refactor(events-page): move listings to a new search results component
PT-1213
